Remove unneeded requirement on gnutls-openssl.
gnutls-openssl is on optional extra of gnutls and is not
installed on some platforms like Fedora since the symbols are
duplicates of libcrypt but not ABI compatible; which can cause
all kinds of crashes and debug headaches.

We were already requiring libgcrypt and it already has an MD5
hash function that is as easy to use as openssl's MD5() function.
So switch to that and make Fedora user's life easier.

Patch originates from Fedora although I tweaked some.